break preview app up into more components
