+++ lib/platform-bible-react/src/preview/components/basics/slider.examples.component.tsx
@@ -1,0 +1,21 @@
+import { useState } from 'react';
+import Slider from '@/components/slider.component';
+
+export default function SliderExamples() {
+  const [sliderValue, setSlider] = useState(3);
+  return (
+    <>
+      Default Shadcn sliders are not themed and always blue 🥺
+      <Slider />
+      <Slider isDisabled />
+      <Slider
+        min={0}
+        max={5}
+        value={sliderValue}
+        onChange={(_e, value) => setSlider(Array.isArray(value) ? value?.[0] : value)}
+      />
+      {/* Wondering in which case the slider would output a number[] as its value 🤷 */}
+      {sliderValue}
+    </>
+  );
+}
