+++ lib/platform-bible-react/src/preview/components/basics/switch.examples.component.tsx
@@ -1,0 +1,16 @@
+import Switch from '@/components/switch.component';
+
+export default function SwitchExamples() {
+  return (
+    <>
+      Default Shadcn switches are not themed and always blue 🥺
+      <div>
+        <Switch />
+        <Switch isDisabled />
+        <Switch isChecked />
+        <Switch isChecked isDisabled />
+        <Switch hasError />
+      </div>
+    </>
+  );
+}
